examples/mnist_classifier/requirements.txt
skeltorch~=1.2.0
torchvision~=0.8.2
